requirements.txt
Flask==2.2.5
line-bot-sdk==1.8.0
pymysql==0.9.2
mysqlclient==2.0.3
python-dotenv==1.0.1
DBUtils==1.3
